Revert "Enhance track management with genre support and route modularization"
