+++ packages/core-storage/src/types.ts
@@ -33,8 +33,6 @@ export type TrackShare = Prisma.TrackShareGetPayload<{
   };
 }>;
 
-export type Genre = Prisma.GenreGetPayload<{}>;
-
 export interface PaginatedResponse<T> {
   items: T[];
   metadata: {
