+++ packages/core-storage/types/index.d.ts
@@ -3,5 +3,5 @@ export { PrismaService } from "./services/prisma";
 export { deleteLocalFile, ensureDirectoryExists, normalizeFilePath, } from "./services/local-storage";
 export { config, type StorageConfig, type DatabaseConfig, type RedisConfig, type SharedConfig, } from "./config";
 export { Prisma, User, Role, SourceFormat, TrackStatus, InviteCode, FeatureFlag, Notification, NotificationType, AudioFileConversionStatus, } from ".prisma/client";
-export { type Track, type Stem, type TrackShare, type Genre, type PaginatedResponse, type PaginationParams, encodeCursor, decodeCursor, } from "./types";
+export { type Track, type Stem, type TrackShare, type PaginatedResponse, type PaginationParams, encodeCursor, decodeCursor, } from "./types";
 export * from "./auth";
